pyproject.toml
[project]
name = "vpa"
version = "0.1.0"
description = "Virtual Package Analysis"
authors = [
    {name = "Matt Hauff", email = "quexington@gmail.com"},
]
dependencies = [
    "click==8.1.7",
    "PyYAML==6.0.1",
    "graphviz",
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
vpa = "vpa.vpa:main"

[tool.setuptools]
package-dir = {"vpa" = "vpa"}

[tool.pdm]
distribution = true
